Move `ColumnSelector` and `Schema` classes into `graph` package
