+++ nvtabular/graph/__init__.py
@@ -17,3 +17,5 @@
 # flake8: noqa
 from .base_operator import BaseOperator
 from .node import Node
+from .schema import ColumnSchema, Schema
+from .selector import ColumnSelector
