+++ nvtabular/ops/graph/concat_columns.py
@@ -14,8 +14,8 @@
 # limitations under the License.
 #
 
-from nvtabular.columns import ColumnSelector
 from nvtabular.dispatch import DataFrameType
+from nvtabular.graph.selector import ColumnSelector
 
 from ..operator import Operator
