+++ tests/unit/columns/test_column_selector.py
@@ -15,7 +15,7 @@
 #
 import pytest
 
-from nvtabular.columns import ColumnSelector
+from nvtabular.graph.selector import ColumnSelector
 from nvtabular.ops import Operator
 from nvtabular.ops.graph.selection import SelectionOp
 from nvtabular.tags import Tags
